// src/types/solicitation.types.ts
export interface Solicitation {
  id: string;
  name_client: string;
  document_client: number;
  contact_client: number;
  address_client: string;
  payment_method: string;
  observations: string;
  pizzas: string;
  order: string;
}
export interface RouteParamsSolicitation {
  id: string;
}

export interface QueryParamsFindMySolicitations {
  id?: string
}